refactor(utils): refactor network utils after migrate to Dagger
