+++ app/src/main/java/com/example/shmr_finance_app_android/data/remote/api/NetworkChecker.kt
@@ -3,7 +3,6 @@ package com.example.shmr_finance_app_android.data.remote.api
 import android.content.Context
 import android.net.ConnectivityManager
 import dagger.Reusable
-import dagger.hilt.android.qualifiers.ApplicationContext
 import javax.inject.Inject
 
 /**
@@ -16,7 +15,7 @@ interface NetworkChecker {
 
 @Reusable
 class NetworkCheckerImpl @Inject constructor(
-    @ApplicationContext private val context: Context
+    private val context: Context
 ) : NetworkChecker {
 
     override fun isNetworkAvailable(): Boolean {
